added DisplayUsersRequestedOrdersCommand to Factory, Filter
