+++ src/main/java/com/my/library/controller/command/constant/commands/AdminCommands.java
@@ -13,4 +13,6 @@ public interface AdminCommands {
     String UNBLOCK_USER = "unblock-user";
 
     String CHANGE_ROLE = "change-user-role";
+
+
 }
