+++ src/main/java/com/my/library/controller/command/constant/commands/LibrarianCommands.java
@@ -6,4 +6,5 @@ public interface LibrarianCommands {
     String RETURN_ORDER = "return-order";
 
     String DISPLAY_READERS = "display-readers";
+    String DISPLAY_REQUESTED_ORDERS = "display-users-requested-orders";
 }
