--- src/main/java/com/my/library/controller/command/impl/librarian/DisplayUsersRequestedOrders.java
@@ -1,24 +1,0 @@
-package com.my.library.controller.command.impl.librarian;
-
-import com.my.library.controller.command.Command;
-import com.my.library.controller.command.CommandResult;
-import com.my.library.exceptions.CommandException;
-import com.my.library.services.OrderService;
-import jakarta.servlet.http.HttpServletRequest;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class DisplayUsersRequestedOrders implements Command {
-    private static final Logger logger = LogManager.getLogger();
-
-    private final OrderService orderService;
-
-    public DisplayUsersRequestedOrders(OrderService orderService) {
-        this.orderService = orderService;
-    }
-
-    @Override
-    public CommandResult execute(HttpServletRequest request) throws CommandException {
-        return null;
-    }
-}
