+++ src/main/java/com/my/library/controller/command/impl/librarian/DisplayUsersRequestedOrdersCommand.java
@@ -1,0 +1,37 @@
+package com.my.library.controller.command.impl.librarian;
+
+import com.my.library.controller.command.Command;
+import com.my.library.controller.command.CommandResult;
+import com.my.library.controller.command.constant.CommandDirection;
+import com.my.library.controller.command.constant.RedirectToPage;
+import com.my.library.controller.command.constant.parameters.Parameters;
+import com.my.library.dto.UserDTO;
+import com.my.library.dto.mapper.UserMapper;
+import com.my.library.entities.User;
+import com.my.library.exceptions.CommandException;
+import com.my.library.exceptions.ServiceException;
+import com.my.library.services.OrderService;
+import com.my.library.utils.Pages;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import java.util.List;
+
+public class DisplayUsersRequestedOrdersCommand implements Command {
+    private static final Logger logger = LogManager.getLogger();
+
+    private final OrderService orderService;
+
+    public DisplayUsersRequestedOrdersCommand(OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @Override
+    public CommandResult execute(HttpServletRequest request) throws CommandException {
+        logger.log(Level.DEBUG, "DisplayUsersRequestedOrdersCommand invoked");
+        return null;
+    }
+}
